requirements.txt
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.5
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.6
msal==1.28.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyarrow==15.0.0
pycparser==2.21
PyJWT==2.8.0
pyperclip==1.8.2
python-dateutil==2.8.2
python-tkdnd==0.2.1
pytube==15.0.0
pytz==2024.1
reportlab==4.1.0
requests==2.31.0
six==1.16.0
tkinterdnd2==0.3.0
ttkwidgets==0.13.0
tzdata==2024.1
urllib3==2.2.1
